# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/drivers": {
      "post": {
        "operationId": "DriversController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      },
      "get": {
        "operationId": "DriversController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      }
    },
    "/drivers/register": {
      "post": {
        "operationId": "DriversController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      }
    },
    "/drivers/{id}": {
      "get": {
        "operationId": "DriversController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      },
      "patch": {
        "operationId": "DriversController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      },
      "delete": {
        "operationId": "DriversController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Drivers"
        ]
      }
    },
...
    "/driver-payout/rickshaw/{rickshawId}": {
      "get": {
        "operationId": "DriverPayoutController_findByRickshaw",
        "parameters": [
          {
            "name": "rickshawId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverPayout"
        ]
      }
    },
    "/driver-payout/{id}": {
      "get": {
        "operationId": "DriverPayoutController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverPayout"
        ]
      },
      "delete": {
        "operationId": "DriverPayoutController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverPayout"
        ]
      }
    },
    "/driver-payout/{id}/pay": {
      "patch": {
        "operationId": "DriverPayoutController_markPaid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverPayout"
        ]
      }
    },
    "/driver-documents": {
      "post": {
        "operationId": "DriverDocumentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "DriverDocuments"
        ]
      },
      "get": {
        "operationId": "DriverDocumentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverDocuments"
        ]
      }
    },
    "/driver-documents/driver/{driverId}": {
      "post": {
        "operationId": "DriverDocumentsController_createForDriver",
        "parameters": [
          {
            "name": "driverId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDriverDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "DriverDocuments"
        ]
      },
      "get": {
        "operationId": "DriverDocumentsController_findByDriver",
        "parameters": [
          {
            "name": "driverId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverDocuments"
        ]
      }
    },
    "/driver-documents/{id}": {
      "get": {
        "operationId": "DriverDocumentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverDocuments"
        ]
      },
      "patch": {
        "operationId": "DriverDocumentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDriverDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverDocuments"
        ]
      },
      "delete": {
        "operationId": "DriverDocumentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverDocuments"
        ]
      }
    },
    "/driver-documents/{id}/verify": {
      "patch": {
        "operationId": "DriverDocumentsController_verify",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyDriverDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "DriverDocuments"
        ]
      }
    },
...
  },
  "info": {
    "title": "AdOnTheGo API",
    "description": "AdOnTheGo Backend APIs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "player-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Enter: ApiKey <your-api-key>"
      }
    },
    "schemas": {
...
      "CreateDriverDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateDriverDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateDriverDocumentDto": {
        "type": "object",
        "properties": {
          "documentType": {
            "type": "string",
            "example": "PUC",
            "enum": [
              "AADHAAR_FRONT",
              "AADHAAR_BACK",
              "PAN_CARD",
              "LICENCE_FRONT",
              "LICENCE_BACK",
              "RC_BOOK",
              "INSURANCE",
              "PUC",
              "PERMIT"
            ]
          },
          "fileUrl": {
            "type": "string",
            "example": "puc.pdf"
          },
          "expiryDate": {
            "type": "string",
            "example": "2027-06-26"
          }
        },
        "required": [
          "documentType",
          "fileUrl"
        ]
      },
      "VerifyDriverDocumentDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateDriverDocumentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}